Implemented a ray that can shoot across the board . It uses coordinates of the entire hexagonal display board and creates a line that goes through each hexagon one at a time until there is no next hexagon.
modified the hexagon class and display board class to get the centre coordinates of the hexagon. Created is therenexthex method which checks if the ray will continue or it has reached the exit. Created a calculate coordinates method. Created a findneresthex method that finds the nearest hexagon from the arrows that are clicked.